+++ src/main/java/com/tecsup/petclinic/services/PetService.java
@@ -31,14 +31,14 @@ public interface PetService {
 	 * @param id
 	 * @throws PetNotFoundException
 	 */
-	void delete(Long id) throws PetNotFoundException;
+	void delete(Integer id) throws PetNotFoundException;
 
 	/**
 	 * 
 	 * @param id
 	 * @return
 	 */
-	Pet findById(long id) throws PetNotFoundException;
+	Pet findById(Integer id) throws PetNotFoundException;
 
 	/**
 	 *
